Cleaned inspector & updated Archive script to allow terrain editing.
Thanks to Frankelstner, the creator of the initial Frostbite mod tools,
we can finally modify terrain!

The original archive script was broken.
It assigned the deleted flag by file size instead of file extension.
This update fixes that and allows modification of the terrain.fbrb
archives.

This means that we can finally have completely new maps instead of only
slightly modified ones.

The next update will hopefully include this.

Other changes:
* Added save button instead of the original tacky-method
* Added minimap preview for map names. This makes it easier to figure
out which map is which
* Cleaned some stuff
* Moved some files
* Fixed some bugs