Fix output of pos values without decimal separator
This time it's not a culture specific issue, everyone is affected by it.
If objects were placed in a way that one or more position values had no decimal places (e.g. 35), it was also written like that to file, but the Frostbite engine expects the position values to be decimal numbers (e.g. 35.0) unless the value is zero, otherwise the position of the object will be messed up ingame. I have added a small fix for it that adds a ".0" to the string if the value is not zero and doesn't contain a dot before writing it to file. That should be accurate enough in this case.